// front/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        "FieldType-field_type.read": {
            readonly id?: number;
            slug: string;
            label: string;
            icon?: string | null;
            description?: string | null;
            /**
             * @default string
             * @enum {string}
             */
            dataType: "string" | "integer" | "float" | "boolean" | "array" | "json" | "date_time";
            component: string;
            assets?: (string | null)[] | null;
            initHook?: string | null;
            configurationSchema?: (string | null)[];
            defaultConfiguration?: (string | null)[];
            /**
             * @default text
             * @enum {string}
             */
            presentationType: "text" | "html" | "date" | "datetime" | "bool" | "image" | "file" | "hidden";
            presentationConfiguration?: (string | null)[];
        };
        "FieldType-field_type.write": {
            slug: string;
            label: string;
            icon?: string | null;
            description?: string | null;
            /**
             * @default string
             * @enum {string}
             */
            dataType: "string" | "integer" | "float" | "boolean" | "array" | "json" | "date_time";
            component: string;
            assets?: (string | null)[] | null;
            initHook?: string | null;
            configurationSchema?: (string | null)[];
            defaultConfiguration?: (string | null)[];
            /**
             * @default text
             * @enum {string}
             */
            presentationType: "text" | "html" | "date" | "datetime" | "bool" | "image" | "file" | "hidden";
            presentationConfiguration?: (string | null)[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}